+++ CHANGELOG.md
@@ -6,12 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/ru/1.0.0/) a
 
 ## [Unreleased]
 
-## [1.5.0] - 2025-01-19
+## [1.5.0] - 2025-01-25
 
 ### Изменено
 
 - Обновление структуры и файлов проекта.
 - Добавление шаблона `Gaa: Project Library`.
+- Добавление шаблона `Gaa: Project Console`.
 
 ## [1.4.1] - 2024-07-19
 
@@ -63,7 +64,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/ru/1.0.0/) a
 - Добавление шаблона `Gaa: Empty Solution`.
 - Добавление шаблона `Gaa: Library Solution`.
 
-[Unreleased]: https://github.com/g-aa/gaa-dotnet-templates/compare/release-v1.4.1...main
+[Unreleased]: https://github.com/g-aa/gaa-dotnet-templates/compare/release-v1.5.0...master
+[1.5.0]: https://github.com/g-aa/gaa-dotnet-templates/compare/release-v1.4.1...release-v1.5.0
 [1.4.1]: https://github.com/g-aa/gaa-dotnet-templates/compare/release-v1.4.0...release-v1.4.1
 [1.4.0]: https://github.com/g-aa/gaa-dotnet-templates/compare/release-v1.3.0...release-v1.4.0
 [1.3.0]: https://github.com/g-aa/gaa-dotnet-templates/compare/release-v1.2.0...release-v1.3.0